rename App component to UrlShortener
