+++ web-client/src/reducers/rootReducer.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function rootReducer () {
+
+}
